Fixed AI under timeWarp by making multiple calls to inGame.performUpdate rather than one with an increased deltaTime

+++ src/de/phbouillon/android/framework/TimeFactorChangeListener.java
@@ -1,5 +1,5 @@
 package de.phbouillon.android.framework;
 
 public interface TimeFactorChangeListener {
-	void timeFactorChanged(float oldTimeFactor, float newTimeFactor);
+	void timeFactorChanged(int oldTimeFactor, int newTimeFactor);
 }
